backend/pyproject.toml
[project]
name = "smart-hydroponic-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "authlib>=1.6.5",
    "fastapi[standard]>=0.118.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "websockets>=15.0.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "bcrypt>=5.0.0",
]
